Sprint 28 -
Atendido - Foi esquecido de mencionar nas sprints anteriores das regras dos usuários.
Alguns tipos de usuários podem editar a atividade após finalizada com a mesma regra da edição do apotamento onde que deverá justificar o motivo da edição e essa justificativa ir para a observação. Ainda não foram informados os cargos que terão essa permissão. Então deixe a regra já implementada porem permitindo todos até sabermos os cargos com esse privilégio.

Lista de Atividades
Atendido -  Linha entre as atividades na lista de atividades para marcar a separação. Caso não seja possivel, voltar ao formato antigo porem com a disposição de colunas como foi feito seguindo o feedback da Gelf.

Atividade
Corrigido - Permitindo finalizar a atividade sem preencher a avaliação da qualidade.
Corrigido -   Permitindo adicionar calibragem após atividade finalizada.

Calibração
Corrigido -   Permitindo salvar calibração mesmo com não conformidade.

Apontamento Diário
Corrigido -   J, P e Q do hora ajudante, hora operador e hora maquina estão sendo cortados.
Corrigido -  A lista de insumos já deve existir só com a opção de acrescentar a quantidade aplicada diretamente na coluna ou por meio de um alert, vide conforme imagem da tela enviada na sprint 25.
Corrigido  -   Permitindo registrar apontamento anterior a data do periodo programado.

Corrigido -  Permitindo salvar a edição de insumo mesmo saindo do alert de justificativa em branco utilizando o botão de voltar do android. Sugestão: Obrigar a colocar justificativa quando clicar em salvar assim como em rendimento. Se tambem houver alteração em rendimento, chamar uma justificativa após outra.

Avaliação Qualidade
Corrigido - Não permitindo a troca de orientação da tela para horizontal.
Corrigido - Manejo com provavel inconsistência de dados. Ex.: com.example

+++ app/src/main/java/com/example/forestsys/calculadora/i/CalculadoraMain.java
@@ -13,7 +13,7 @@ import com.example.forestsys.R;
 public class CalculadoraMain extends AppCompatActivity {
 
     private TextView e1, e2;
-    private int contador = 0;
+    private long contador = 0;
     private String expressao = "";
     private String texto = "";
     private Double resultado = 0.0;
